Have make corrections as per review
