+++ package.json
@@ -4,8 +4,9 @@
 	"description": "",
 	"main": "./dist/app.js",
 	"scripts": {
-		"build": "tsc",
+		"build": "rimraf ./build && tsc",
 		"start": "npm run build && node ./dist/app.js",
+		"start:dev": "nodemon src/app.ts",
 		"test": "echo \"Error: no test specified\" && exit 1"
 	},
 	"repository": {
@@ -28,6 +29,7 @@
 		"@types/swagger-jsdoc": "^6.0.1",
 		"@types/swagger-ui-express": "^4.1.4",
 		"nodemon": "^3.0.1",
+		"rimraf": "^5.0.5",
 		"ts-node": "^10.9.1",
 		"typescript": "^5.2.2"
 	}
