--- src/swagger.ts
@@ -1,17 +1,0 @@
-import swaggerJSDoc from "swagger-jsdoc";
-
-const swaggerOptions = {
-	definition: {
-		openapi: "3.0.0",
-		info: {
-			title: "node.js_course_practice",
-			version: "1.0.0",
-			description: "Documentation for my Node.js API",
-		},
-	},
-	apis: ["./app.js"],
-};
-
-const swaggerSpec = swaggerJSDoc(swaggerOptions);
-
-export default swaggerSpec;
